healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>

Former-commit-id: c932b9758e7da2d4e6676656d3ee888acbdb3a22
Former-commit-id: 900e7bc23bf86eefd9b9310ab91c21cb75fc7043
Former-commit-id: 8d45bd54c048bd85ec5466af5b944f81d0328b45