remount: Fix failure for system-as-root
Since commit 1a41cac71999 [formerly 1a41cac71999 [formerly 1a41cac71999 [formerly 5ad7b3cbc5c1]]] ("Try to remount mounted points only"), the
system-as-root cannot be remounted successfully when execute command
"adb remount".  This is because the mount point "/system" cannot be
found and directly bails out with failure.

Add an extra checking for the mount point "/system" for the
system-as-root case, takes it as a found entry; thus the function can
continue to run and fix up to remount root.

Suggested-by: David Anderson <dvander@google.com>
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Change-Id: Ia936c9d97bed951184813a087c70fe591cb33fe0

Former-commit-id: c411894698009028853067254506330f549da32d
Former-commit-id: 3336eae47eb119857c322684a9dd8a4339a41e65
Former-commit-id: 06dcf174e92386aec2ec790243100e1037565bb8